pyproject.toml
[tool.poetry]
name = "paperless-asn-labels"
version = "0.1.0"
description = "Creates archive serial number (ASN) label sheets"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fpdf2 = "^2.7.8"
segno = "^1.6.1"
click = "^8.1.7"

[tool.poetry.scripts]
paperless-asn-labels = "paperless_asn_labels.main:generate"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# pycodestyle (E)
# pyflakes (F)
# pydocstyle (D)
# pyupgrade (UP)
# pylint (PL)
# enforce use of pathlib (PTH)
# disallow access to private members (SLF)
# stricter return type handling (RET)
# exception linting (RSE)
# naming conventions (N)
lint.select = ["E", "F", "D", "UP", "PL", "PTH", "SLF", "RET", "RSE", "N"]

# Pylint refactor (PLR), pylint complexity warnings (PLC0414)
lint.ignore = ["PLR", "PLC0414"]

line-length = 120
target-version = "py310"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
